requirements.txt
crewai==0.98.0
crewai-tools==0.32.1
langchain-community==0.3.15
langchain-groq==0.2.3
python-dotenv==1.0.1